Move ANDRDOID_SDK_INT to JvmPlatform file
